move logic get, create users to user-controller
